Move Generator to ./parts
