+++ packages/openapi-codegen-client-fresha/src/command.ts
@@ -4,7 +4,7 @@ import { createLogger } from '@fresha/openapi-codegen-utils';
 import { OpenAPIReader } from '@fresha/openapi-model/build/3.0.3';
 import { Project } from 'ts-morph';
 
-import { Generator } from './Generator';
+import { Generator } from './parts';
 
 import type { Argv, ArgumentsCamelCase } from 'yargs';
