--- packages/openapi-codegen-client-fresha/src/index.test.ts
@@ -1,3 +1,0 @@
-test('dummy', () => {
-  expect(2 + 2).toBe(4);
-});
