+++ packages/openapi-codegen-client-fresha/src/index.ts
@@ -1,1 +1,1 @@
-export { Generator } from './Generator';
+export { Generator } from './parts';
